delete diagnostic and quiz
